requirements.txt
embedchain
langchain
langchain_openai
wikipedia
arxiv
newspaper3k
pypdf
lxml[html_clean]
duckduckgo-search
pycountry